This file is a python script
This script is a function that allows a user to input arguments for start, stop and step in a range. unlike the default range, it allows the inclusive. In other words, it includes the terminal value of the range. One more unique function of this script is that it sums up the range.